requirements.txt
# Minimal working requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
plotly>=5.17.0
requests>=2.32.0
python-multipart>=0.0.12

# Google AI - Let it manage its own dependencies

google-generativeai>=0.8.5
langchain-google-genai==2.0.10
# Essential AI libraries
langchain
langchain-community
crewai
pydantic-settings

faiss-cpu
gitpython==3.1.37